Move transport layer into functions/lambda
