+++ functions/lambda/main.go
@@ -5,10 +5,14 @@ import (
 
 	"github.com/aws/aws-lambda-go/lambda"
 	"github.com/meetnearme/api/functions/lambda/handlers"
-	transport "github.com/meetnearme/api/internal/transport/lambda"
+	"github.com/meetnearme/api/functions/lambda/transport"
 )
 
 var router *transport.Router
+
+// https://8hpnqnaevi.execute-api.us-east-1.amazonaws.com/
+// https://github.com/raphael-p/beango-messenger/blob/master/server/server.go
+// https://raphael-p.medium.com/a-guide-to-making-a-go-web-server-without-a-framework-1439a965f2b1
 
 func init() {
 	router = transport.NewRouter()
